Traps added
