--- Assets/Simple Helvetica/Editor/SimpleHelvetica.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 6fcd2d0ed406e4e0093d1f6f1bf7c603
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
